Javadoc (#239)
new javadoc parsing

+++ src/main/kotlin/javabot/model/ApiEvent.kt
@@ -3,13 +3,15 @@ package javabot.model
 import javabot.JavabotConfig
 import javabot.dao.AdminDao
 import javabot.dao.ApiDao
-import javabot.model.javadoc.JavadocApi
 import javabot.javadoc.JavadocParser
 import javabot.model.EventType.DELETE
 import javabot.model.EventType.RELOAD
+import javabot.model.javadoc.JavadocApi
 import org.bson.types.ObjectId
 import org.mongodb.morphia.annotations.Entity
+import org.mongodb.morphia.annotations.Reference
 import org.mongodb.morphia.annotations.Transient
+import org.slf4j.LoggerFactory
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
@@ -26,52 +28,33 @@ import javax.inject.Inject
 class ApiEvent : AdminEvent {
 
     companion object {
-        fun locateJDK(): URI {
-            var property = System.getProperty("java.home")
-            if (property.endsWith("/jre")) {
-                property = property.dropLast(4)
-            }
-            var home = File(property)
-            var file = File(property, "src.zip")
-
-            while (!file.exists()) {
-                home = home.parentFile
-                file = File(home, "src.zip")
-            }
-            return file.toURI()
+        fun add(requestedBy: String, api: JavadocApi): ApiEvent {
+            return ApiEvent(requestedBy, api)
         }
         
-        fun add(requestedBy: String, name: String, groupId: String = "",
-                artifactId: String = "", version: String = ""): ApiEvent {
-            return ApiEvent(requestedBy, name, groupId, artifactId, version)
-        }
-        
-        fun drop(requestedBy: String, id: ObjectId) : ApiEvent {
-            return ApiEvent(requestedBy, DELETE, id)
+        fun drop(requestedBy: String, api: JavadocApi) : ApiEvent {
+            return ApiEvent(requestedBy, DELETE, api)
         }
 
         fun drop(requestedBy: String, name: String) : ApiEvent {
             return ApiEvent(requestedBy, DELETE, name)
         }
 
-        fun reload(requestedBy: String, id: ObjectId) : ApiEvent {
-            return ApiEvent(requestedBy, RELOAD, id)
+        fun reload(requestedBy: String, api: JavadocApi) : ApiEvent {
+            return ApiEvent(requestedBy, RELOAD, api)
         }
 
         fun reload(requestedBy: String, name: String) : ApiEvent {
             return ApiEvent(requestedBy, RELOAD, name)
         }
-    }
 
-    var apiId: ObjectId? = null
+        private val LOG = LoggerFactory.getLogger(ApiEvent::class.java)
+    }
 
     lateinit var name: String
 
-    lateinit var groupId: String
-
-    lateinit var artifactId: String
-
-    lateinit var version: String
+    @Reference(idOnly = true, ignoreMissing = true)
+    var api: JavadocApi? = null
 
     @Inject
     @Transient
@@ -89,20 +72,16 @@ class ApiEvent : AdminEvent {
     @Transient
     lateinit var adminDao: AdminDao
 
-    constructor() {
-    }
+    constructor()
 
-    private constructor(requestedBy: String, name: String, groupId: String, artifactId: String, version: String) :
+    private constructor(requestedBy: String, api: JavadocApi) :
     super(requestedBy, EventType.ADD) {
         this.requestedBy = requestedBy
-        this.name = name
-        this.groupId = groupId
-        this.artifactId = artifactId
-        this.version = version
+        this.api = api
     }
 
-    private constructor(requestedBy: String, type: EventType, apiId: ObjectId?) : super(requestedBy, type) {
-        this.apiId = apiId
+    private constructor(requestedBy: String, type: EventType, api: JavadocApi) : super(requestedBy, type) {
+        this.api = api
     }
 
     private constructor(requestedBy: String, type: EventType, name: String) : super(requestedBy, type) {
@@ -115,58 +94,41 @@ class ApiEvent : AdminEvent {
     }
 
     override fun delete() {
-        var api = findApi()
-        if (api != null) {
-            File("javadoc/${api.name}/${api.version}/").deleteTree()
-            apiDao.delete(api)
-        }
-    }
-
-    private fun findApi(): JavadocApi? {
-        var api = apiDao.find(apiId)
-        if (api == null) {
+        if(api == null) {
             api = apiDao.find(name)
         }
-        return api
+        api?.let {
+            File("javadoc/${it.name}/${it.version}/").deleteTree()
+            apiDao.delete(it)
+            it.id = ObjectId()
+        }
     }
 
+    private fun findApi()= api ?: apiDao.find(name)
+
     override fun add() {
-        val api = JavadocApi(name, config.url(), groupId, artifactId, version)
-        process(api)
+        api?.let {
+            apiDao.save(it)
+            val admin = adminDao.getAdmin(JavabotUser(requestedBy, requestedBy, ""))
+            if (admin != null) {
+                val user = JavabotUser(admin.ircName, admin.emailAddress, admin.hostName)
+
+                parser.parse(it, object : StringWriter() {
+                    override fun write(line: String) {
+                        bot.privateMessageUser(user, line)
+                        LOG.debug(line)
+                    }
+                })
+            }
+        }
     }
 
     override fun reload() {
         val api = findApi()
         if (api != null) {
-            name = api.name
-            groupId = api.groupId
-            artifactId = api.artifactId
-            version = if (name == "JDK") System.getProperty("java.version") else api.version
-//            apiDao.delete(apiId)
-//            api.id = ObjectId()
-//            apiDao.save(api)
             delete()
-            process(api)
+            add()
         }
-    }
-
-    private fun process(api: JavadocApi) {
-        apiDao.save(api)
-        val downloadUrl = if (name == "JDK") locateJDK() else buildMavenUrl()
-        val admin = adminDao.getAdmin(JavabotUser(requestedBy, requestedBy, ""))
-        if (admin != null) {
-            val user = JavabotUser(admin.ircName, admin.emailAddress, admin.hostName)
-
-            parser.parse(api, downloadUrl.downloadZip(),
-                    object : StringWriter() {
-                        override fun write(line: String) = bot.privateMessageUser(user, line)
-                    })
-        }
-    }
-
-    private fun buildMavenUrl(): URI {
-        return URI("https://repo1.maven.org/maven2/${groupId.replace(".", "/")}/${artifactId}" +
-                "/${version}/${artifactId}-${version}-sources.jar")
     }
 
     override fun toString(): String {
@@ -175,10 +137,14 @@ class ApiEvent : AdminEvent {
 }
 
 fun URI.downloadZip(): File {
-    val file = File.createTempFile("javadoc-", ".zip")
-    FileOutputStream(file).use { outputStream ->
-        this.toURL().openStream().use { inputStream ->
-            outputStream.write(inputStream.readBytes())
+    val downloads = File("downloads")
+    downloads.mkdir()
+    val file = File(downloads, path.substringAfterLast("/"))
+    if(!file.exists()) {
+        FileOutputStream(file).use { outputStream ->
+            this.toURL().openStream().use { inputStream ->
+                outputStream.write(inputStream.readBytes())
+            }
         }
     }
     return file
